[refactor/72-task] 검증로직 추가
[refactor/72-task] 검증로직 추가
