+++ ideac-core/src/main/java/depth/main/ideac/domain/auth/dto/request/SignUpReq.java
@@ -1,6 +1,7 @@
 package depth.main.ideac.domain.auth.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
+import jakarta.persistence.Column;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
 import jakarta.validation.constraints.Size;
@@ -13,6 +14,7 @@ public class SignUpReq {
     private String idEmail;
 
     @Schema( type = "string", example = "홍길동", description="이름")
+    @Size(min = 2, message = "2자 이상 입력해주세요")
     @NotBlank
     private String name;
 
@@ -33,6 +35,7 @@ public class SignUpReq {
 
     @Schema( type = "string", example = "01012341234", description="휴대폰번호")
     @NotBlank
+    @Pattern(regexp = "^01([0|1|6|7|8|9])?([0-9]{3,4})?([0-9]{4})$", message = "번호를 정확하게 입력해주세요")
     private String phoneNumber;
 
     @Schema( type = "string", example = "depth", description="소속 동아리")
