+++ ideac-user/src/main/java/depth/main/ideac/domain/idea_post/dto/request/RegisterIdeaReq.java
@@ -12,6 +12,7 @@ public class RegisterIdeaReq {
     @NotBlank(message = "내용이 입력되지 않았습니다.")
     @Size(max = 50, message = "간단 설명은 최대 50자까지 입력 가능합니다.")
     private String simpleDescription;
+    @NotBlank
     private String keyword;
     @NotBlank(message = "내용이 입력되지 않았습니다.")
     private String detailedDescription;
